+++ modules/contracts/apps/subscription-contributors/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "abstract-subscription"
+name = "abstract-subscription-contributors"
 version = { workspace = true }
 authors = { workspace = true }
 edition = { workspace = true }
@@ -13,41 +13,38 @@ crate-type = ["cdylib", "rlib"]
 name = "schema"
 required-features = ["schema"]
 
-[[example]]
-name = "deploy"
-required-features = ["interface"]
-
 [features]
 default = ["export"]
 export = []
-interface = ["export", "dep:abstract-interface", "dep:cw-orch"]
+interface = [
+    "export",
+    "dep:abstract-interface",
+    "dep:cw-orch",
+    "abstract-subscription-interface/interface",
+]
 schema = ["abstract-app/schema"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 cosmwasm-std = { workspace = true }
 cw-controllers = { workspace = true }
 cw-storage-plus = { workspace = true }
 cosmwasm-schema = { workspace = true }
-cw2 = { workspace = true }
-cw20 = { workspace = true }
-schemars = { workspace = true }
-serde = { workspace = true }
 thiserror = { workspace = true }
 cw-asset = { workspace = true }
-semver = { workspace = true }
 
-abstract-sdk = { workspace = true }
 abstract-app = { workspace = true }
 abstract-core = { workspace = true }
-cw-orch = { workspace = true, optional = true }
-
+abstract-sdk = { workspace = true }
 abstract-interface = { workspace = true, optional = true }
-
 abstract-subscription-interface = { workspace = true }
 
+cw-orch = { workspace = true, optional = true }
+
 [dev-dependencies]
-cw20 = { workspace = true, features = ["interface"] }
-# Deploy
+clap = { workspace = true }
+semver = { workspace = true }
 dotenv = "0.15.0"
 env_logger = "0.10.0"
-clap = { workspace = true }
+speculoos = { workspace = true }
