+++ modules/contracts/apps/subscription-contributors/rustfmt.toml
@@ -1,4 +1,4 @@
 # stable
-newline_style = "unix"
+newline_style = "Unix"
 hard_tabs = false
 tab_spaces = 4
