+++ modules/contracts/apps/subscription-contributors/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod contract;
+mod handlers;
+pub use abstract_subscription_interface::msg::contributors as msg;
+mod replies;
+pub use abstract_subscription_interface::state::contributors as state;
+
+#[cfg(feature = "interface")]
+pub use contract::interface::AppInterface;
+#[cfg(feature = "interface")]
+pub use msg::{AppExecuteMsgFns, AppQueryMsgFns};
