+++ modules/contracts/apps/subscription-contributors/src/replies/mod.rs
@@ -1,0 +1,5 @@
+mod instantiate;
+
+pub use instantiate::instantiate_reply;
+
+pub const INSTANTIATE_REPLY_ID: u64 = 1u64;
