+++ modules/contracts/apps/subscription/examples/deploy.rs
@@ -1,3 +1,6 @@
+use abstract_interface::*;
+use abstract_subscription::contract::interface::SubscriptionApp;
+use abstract_subscription_interface::SUBSCRIPTION_ID;
 use clap::Parser;
 use cw_orch::{
     anyhow,
@@ -5,9 +8,6 @@ use cw_orch::{
     prelude::{networks::parse_network, DaemonBuilder},
     tokio::runtime::Runtime,
 };
-
-use abstract_interface::*;
-use abstract_subscription::{contract::interface::SubscriptionApp, SUBSCRIPTION};
 use dotenv::dotenv;
 use semver::Version;
 
@@ -21,7 +21,7 @@ fn deploy_subscription(networks: Vec<ChainInfo>) -> anyhow::Result<()> {
             .chain(network)
             .handle(rt.handle())
             .build()?;
-        let subscription_app = SubscriptionApp::new(SUBSCRIPTION, chain);
+        let subscription_app = SubscriptionApp::new(SUBSCRIPTION_ID, chain);
         subscription_app.deploy(version.clone())?;
     }
     Ok(())
