+++ modules/packages/apps/subscription/src/error/mod.rs
@@ -1,0 +1,2 @@
+pub mod contributors;
+pub mod subscription;
