+++ modules/packages/apps/subscription/src/msg/subscription.rs
@@ -28,22 +28,27 @@
 //! * The total income of the system is shared between the DAO and the contributors. See [`ContributionConfig`].
 //! * (optional) Token emissions to contributor (and users) are dynamically set based on the protocol's income. Meaning that the token emissions will rise if demand/income falls and vice-versa.
 
-use crate::contract::SubscriptionApp;
-use crate::state::UncheckedEmissionType;
-use crate::state::{
-    Compensation, ContributionState, ContributorsConfig, Subscriber, SubscribersConfig,
-    SubscriptionState,
-};
+use crate::state::subscription::UncheckedEmissionType;
+use crate::state::subscription::{Subscriber, SubscribersConfig, SubscriptionState};
+use abstract_core::app;
 use abstract_core::objects::AccountId;
 use cosmwasm_schema::QueryResponses;
 use cosmwasm_std::{Decimal, Uint128, Uint64};
 use cw_asset::{Asset, AssetInfoUnchecked};
 
-abstract_app::app_msg_types!(
-    SubscriptionApp,
-    SubscriptionExecuteMsg,
-    SubscriptionQueryMsg
-);
+use super::contributors::ContributorsInstantiateMsg;
+
+/// Top-level Abstract App execute message. This is the message that is passed to the `execute` entrypoint of the smart-contract.
+pub type ExecuteMsg = app::ExecuteMsg<SubscriptionExecuteMsg>;
+/// Top-level Abstract App instantiate message. This is the message that is passed to the `instantiate` entrypoint of the smart-contract.
+pub type InstantiateMsg = app::InstantiateMsg<SubscribersInstantiateMsg>;
+/// Top-level Abstract App query message. This is the message that is passed to the `query` entrypoint of the smart-contract.
+pub type QueryMsg = app::QueryMsg<SubscriptionQueryMsg>;
+/// Top-level Abstract App migrate message. This is the message that is passed to the `query` entrypoint of the smart-contract.
+pub type MigrateMsg = app::MigrateMsg<AppMigrateMsg>;
+
+impl app::AppExecuteMsg for SubscriptionExecuteMsg {}
+impl app::AppQueryMsg for SubscriptionQueryMsg {}
 
 /// Subscription migration message
 #[cosmwasm_schema::cw_serde]
@@ -69,25 +74,8 @@ pub struct SubscribersInstantiateMsg {
     pub subscription_cost_per_block: Decimal,
     /// Subscription emissions per block
     pub subscription_per_block_emissions: UncheckedEmissionType,
-}
-
-/// Subscribers instantiation message
-#[cosmwasm_schema::cw_serde]
-pub struct ContributorsInstantiateMsg {
-    /// Percentage of income that is redirected to the protocol
-    pub protocol_income_share: Decimal,
-    /// Percentage of emissions allocated to users
-    pub emission_user_share: Decimal,
-    /// Max emissions (when income = 0) = max_emissions_multiple * floor_emissions
-    pub max_emissions_multiple: Decimal,
-    /// Emissions amplification factor in inverse emissions <-> target equation
-    pub emissions_amp_factor: Uint128,
-    /// Emissions offset factor in inverse emissions <-> target equation
-    pub emissions_offset: Uint128,
     /// How often update income average
     pub income_averaging_period: Uint64,
-    /// token: TODO
-    pub token_info: AssetInfoUnchecked,
 }
 
 /// App execution messages
@@ -109,30 +97,9 @@ pub enum SubscriptionExecuteMsg {
         /// List of inactive accounts to move to the `DORMANT_SUBSCRIBERS` list
         os_ids: Vec<AccountId>,
     },
-    /// Claim the compensation for contributor
-    ClaimCompensation {
-        /// Abstract account id of contributor
-        os_id: AccountId,
-    },
     /// Claim the emissions for subscriber
     ClaimEmissions {
         /// Abstract account id of subscriber
-        os_id: AccountId,
-    },
-    /// Update/add the contributor config
-    UpdateContributor {
-        /// Abstract account id of contributor
-        os_id: AccountId,
-        /// Base amount payment per block
-        base_per_block: Option<Decimal>,
-        /// Weight of the contributor
-        weight: Option<Uint64>,
-        /// Block id when "contract" with this contributor expires
-        expiration_block: Option<Uint64>,
-    },
-    /// Remove the contributor
-    RemoveContributor {
-        /// Abstract account id of contributor
         os_id: AccountId,
     },
     /// Update config of subscription
@@ -144,21 +111,6 @@ pub enum SubscriptionExecuteMsg {
         /// new subscription_cost_per_block
         subscription_cost_per_block: Option<Decimal>,
         // TODO?: subscription_per_block_emissions
-    },
-    /// Update config of contributors
-    UpdateContributionConfig {
-        /// New ercentage of income that is redirected to the protocol
-        protocol_income_share: Option<Decimal>,
-        /// New ercentage of emissions allocated to users
-        emission_user_share: Option<Decimal>,
-        /// New max emissions (when income = 0) = max_emissions_multiple * floor_emissions
-        max_emissions_multiple: Option<Decimal>,
-        /// New emissions amplification factor in inverse emissions <-> target equation
-        emissions_amp_factor: Option<Uint128>,
-        /// New emissions offset factor in inverse emissions <-> target equation
-        emissions_offset: Option<Uint128>,
-        /// Change project token
-        project_token_info: Option<AssetInfoUnchecked>,
     },
 }
 
@@ -183,12 +135,6 @@ pub enum SubscriptionQueryMsg {
         /// Abstract Account Id of subscriber  
         os_id: AccountId,
     },
-    /// Get state of the contributor
-    #[returns(ContributorStateResponse)]
-    ContributorState {
-        /// Abstract Account Id of contributor
-        os_id: AccountId,
-    },
 }
 
 /// Cw20 hook message
@@ -204,8 +150,6 @@ pub enum DepositHookMsg {
 /// Query response for [`SubscriptionQueryMsg::Config`]
 #[cosmwasm_schema::cw_serde]
 pub struct ConfigResponse {
-    /// Config for the contributors
-    pub contribution: ContributorsConfig,
     /// Config for the subscribers
     pub subscription: SubscribersConfig,
 }
@@ -213,8 +157,6 @@ pub struct ConfigResponse {
 /// Query response for [`SubscriptionQueryMsg::State`]
 #[cosmwasm_schema::cw_serde]
 pub struct StateResponse {
-    /// State of contributors
-    pub contribution: ContributionState,
     /// State of subscribers
     pub subscription: SubscriptionState,
 }
@@ -233,11 +175,4 @@ pub struct SubscriberStateResponse {
     pub currently_subscribed: bool,
     /// State of the subscription
     pub subscriber_details: Subscriber,
-}
-
-/// Query response for [`SubscriptionQueryMsg::ContributorState`]
-#[cosmwasm_schema::cw_serde]
-pub struct ContributorStateResponse {
-    /// Compensation details for contributors
-    pub compensation: Compensation,
 }
